object editing finished
You can now edit the values for each node by changing the values in the
text boxes and clicking a button. To get this to work put the ui file
the folder of your solution, change the location of myglwidget.h in the
generated file to the one in the github folder, and do the same for the
new inputLine object I had to make to get the user input to work. I also
added an isMesh property to the base geometry which will be used later
when we put in the subdivide feature.